# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Production backend for UnityWorks Vision AI. Owns the Vision OS perception platform and exposes it through an authorized API.",
    "title": "UnityWorks Vision AI",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cameras": {
      "get": {
        "operationId": "list_cameras_api_v1_cameras_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Cameras Api V1 Cameras Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Cameras",
        "tags": [
          "product"
        ]
      },
      "post": {
        "description": "Register a camera. Created **disabled** \u2014 enabling is a separate act.",
        "operationId": "create_camera_api_v1_cameras_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Camera Api V1 Cameras Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Camera",
        "tags": [
          "product"
        ]
      }
    },
...
    "/api/v1/cameras/{camera_key}": {
      "delete": {
        "description": "Remove a camera and destroy its observation partition.\n\n### Its own permission, not `MANAGE_CAMERAS`\n\nGated on `RETIRE_CAMERAS`. Everything this docstring goes on to say about\ndestroying records is the argument for the separation: an account that may\nadd a camera and correct its name has not thereby been trusted to close an\nobservation partition, and no amount of renaming undoes this.\n\n### Two audit rows, always\n\nDeleting a camera is a configuration change *and* a destruction of records\nabout people at work, so it writes both `camera.deleted` and\n`observation.truncated`. A deletion that left only the first would be a\nrecord of the change with no record of the data it destroyed \u2014 and the\nobservation sweep's own discipline is that every truncation is provable.\n\n### The refusal is the point\n\nIf this process cannot reach a durable observation log, the camera is not\ndeleted and the caller is told why. Deleting it there would orphan the\npartition: retention enumerates partitions from the camera table, so a\ndeleted row means nothing ever sweeps those observations again.",
        "operationId": "delete_camera_api_v1_cameras__camera_key__delete",
        "parameters": [
          {
            "in": "path",
            "name": "camera_key",
            "required": true,
            "schema": {
              "title": "Camera Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Camera Api V1 Cameras  Camera Key  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Camera",
        "tags": [
          "product"
        ]
      },
      "patch": {
        "operationId": "update_camera_api_v1_cameras__camera_key__patch",
        "parameters": [
          {
            "in": "path",
            "name": "camera_key",
            "required": true,
            "schema": {
              "title": "Camera Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Camera Api V1 Cameras  Camera Key  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Camera",
        "tags": [
          "product"
        ]
      }
    },
...
    "/api/v1/wall/cameras/{camera_id}": {
      "get": {
        "description": "One camera, for the detail view. Carries no credential and no URL.",
        "operationId": "camera_detail_api_v1_wall_cameras__camera_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "camera_id",
            "required": true,
            "schema": {
              "title": "Camera Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Camera Detail Api V1 Wall Cameras  Camera Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Camera Detail",
        "tags": [
          "camera-wall"
        ]
      }
    },
    "/api/v1/wall/cameras/{camera_id}/stream.mjpg": {
      "get": {
        "description": "`multipart/x-mixed-replace` JPEG. Authorized by ticket, not by header.\n\nDeliberately **not** behind `requires(...)`: an `<img>` tag cannot send an\nAuthorization header, so this route authorizes on the signed ticket the\ncaller obtained from an authenticated request. The ticket names the tenant,\nthe camera and the subject, and all three are checked here.",
        "operationId": "stream_camera_api_v1_wall_cameras__camera_id__stream_mjpg_get",
        "parameters": [
          {
            "in": "path",
            "name": "camera_id",
            "required": true,
            "schema": {
              "title": "Camera Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ticket",
            "required": true,
            "schema": {
              "title": "Ticket",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant",
            "required": true,
            "schema": {
              "title": "Tenant",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subject",
            "required": true,
            "schema": {
              "title": "Subject",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fps",
            "required": false,
            "schema": {
              "default": 4.0,
              "maximum": 25.0,
              "minimum": 0.5,
              "title": "Fps",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Camera",
        "tags": [
          "camera-wall"
        ]
      }
    },
    "/api/v1/wall/cameras/{camera_id}/ticket": {
      "post": {
        "description": "Exchange an authenticated session for a one-camera viewing ticket.",
        "operationId": "issue_ticket_api_v1_wall_cameras__camera_id__ticket_post",
        "parameters": [
          {
            "in": "path",
            "name": "camera_id",
            "required": true,
            "schema": {
              "title": "Camera Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Issue Ticket Api V1 Wall Cameras  Camera Id  Ticket Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Ticket",
        "tags": [
          "camera-wall"
        ]
      }
    },
...
  }
}